[tune] Remove TrialExecutor base class
